requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data quality and validation
great-expectations>=0.17.0
jsonschema>=4.0.0

# Analytics and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Monitoring and logging
psutil>=5.8.0
structlog>=23.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# CI/CD
pre-commit>=3.0.0
